+++ app/src/main/java/com/example/habittracker/AddNewHabitActivity.java
@@ -14,7 +14,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.SwitchCompat;
 
 import com.example.habittracker.controllers.HabitListController;
-import com.example.habittracker.models.Habit;
+import com.example.habittracker.models.Habit.Habit;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
