+++ app/src/main/java/com/example/habittracker/UserProfileActivity.java
@@ -1,14 +1,9 @@
 package com.example.habittracker;
 
-import android.app.Fragment;
-import android.app.FragmentManager;
-import android.app.FragmentTransaction;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.ImageView;
@@ -21,10 +16,10 @@ import androidx.appcompat.app.AppCompatActivity;
 import com.example.habittracker.adapters.UserHabitListAdapter;
 import com.example.habittracker.controllers.HabitListController;
 import com.example.habittracker.controllers.SocialController;
-import com.example.habittracker.models.Habit;
-import com.example.habittracker.models.HabitList;
-import com.example.habittracker.models.Request;
-import com.example.habittracker.models.User;
+import com.example.habittracker.models.Habit.Habit;
+import com.example.habittracker.models.Habit.HabitList;
+import com.example.habittracker.models.Request.Request;
+import com.example.habittracker.models.User.User;
 import com.example.habittracker.ui.follow.FollowersActivity;
 import com.example.habittracker.ui.follow.FollowingActivity;
 import com.google.firebase.firestore.FirebaseFirestore;
