+++ app/src/main/java/com/example/habittracker/ViewHabitEventActivity.java
@@ -6,19 +6,13 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.example.habittracker.adapters.HabitEventListAdapter;
-import com.example.habittracker.controllers.HabitEventsController;
-import com.example.habittracker.interfaces.OnHabitEventsRetrieved;
-import com.example.habittracker.models.DownloadImageTask;
-import com.example.habittracker.models.HabitEvent;
-import com.example.habittracker.models.HabitEventList;
-import com.example.habittracker.models.HabitList;
+import com.example.habittracker.models.HabitEvent.HabitEvent;
+import com.example.habittracker.models.Habit.HabitList;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.FirebaseFirestore;
