+++ app/src/main/java/com/example/habittracker/controllers/HabitEventsController.java
@@ -3,28 +3,23 @@ package com.example.habittracker.controllers;
 import android.os.Build;
 import android.util.Log;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.RequiresApi;
 import com.example.habittracker.interfaces.OnHabitEventDeleted;
 import com.example.habittracker.interfaces.OnHabitEventsRetrieved;
-import com.example.habittracker.models.Habit;
-import com.example.habittracker.models.HabitEvent;
-import com.example.habittracker.models.HabitEventList;
-import com.example.habittracker.models.HabitList;
-import com.example.habittracker.interfaces.OnHabitListRetrieved;
+import com.example.habittracker.models.Habit.Habit;
+import com.example.habittracker.models.HabitEvent.HabitEvent;
+import com.example.habittracker.models.HabitEvent.HabitEventList;
+import com.example.habittracker.models.Habit.HabitList;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.Timestamp;
 import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FieldValue;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 import com.google.firebase.firestore.SetOptions;
 
-import java.sql.Time;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.ZoneOffset;
 import java.util.Date;
