+++ app/src/main/java/com/example/habittracker/models/Habit/HabitList.java
@@ -1,4 +1,4 @@
-package com.example.habittracker.models;
+package com.example.habittracker.models.Habit;
 
 import android.os.Build;
