+++ app/src/main/java/com/example/habittracker/models/HabitEvent/HabitEventList.java
@@ -1,4 +1,4 @@
-package com.example.habittracker.models;
+package com.example.habittracker.models.HabitEvent;
 
 import android.os.Build;
