+++ app/src/main/java/com/example/habittracker/ui/follow/FollowersActivity.java
@@ -3,8 +3,6 @@ package com.example.habittracker.ui.follow;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.View;
-import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.ListView;
@@ -16,15 +14,14 @@ import com.example.habittracker.UserProfileActivity;
 import com.example.habittracker.adapters.FollowersAdapter;
 import com.example.habittracker.controllers.UsersListController;
 import com.example.habittracker.models.Follow.Follow;
-import com.example.habittracker.models.User;
-import com.example.habittracker.models.UsersList;
+import com.example.habittracker.models.User.User;
+import com.example.habittracker.models.User.UsersList;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.FieldPath;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 
 import java.util.ArrayList;
-import java.util.Map;
 
 public class FollowersActivity extends AppCompatActivity {
