+++ app/src/main/java/com/example/habittracker/ui/habitEvents/HabitEventFragment.java
@@ -30,9 +30,9 @@ import com.example.habittracker.Plugins.NonScrollListView;
 import com.example.habittracker.R;
 import com.example.habittracker.ViewHabitEventActivity;
 import com.example.habittracker.adapters.HabitEventListAdapter;
-import com.example.habittracker.models.HabitEvent;
-import com.example.habittracker.models.HabitEventList;
-import com.example.habittracker.models.HabitList;
+import com.example.habittracker.models.HabitEvent.HabitEvent;
+import com.example.habittracker.models.HabitEvent.HabitEventList;
+import com.example.habittracker.models.Habit.HabitList;
 import com.example.habittracker.controllers.HabitEventsController;
 import com.example.habittracker.controllers.HabitListController;
 import com.example.habittracker.interfaces.OnHabitEventsRetrieved;
